increase-code-coverage-one Fix directory hierarchy problem.
