+++ src/test/java/org/apache/commons/imaging/formats/bmp/BmpWriterRgbTest.java
@@ -1,0 +1,35 @@
+package org.apache.commons.imaging.formats.bmp;
+
+import org.junit.Test;
+
+import java.awt.image.BufferedImage;
+import java.util.Arrays;
+
+import static org.junit.Assert.assertEquals;
+
+/**
+ * Unit tests for class {@link BmpWriterRgb}.
+ *
+ * @date 13.07.2017
+ * @see BmpWriterRgb
+ *
+ **/
+public class BmpWriterRgbTest{
+
+
+    @Test
+    public void testGetImageData() {
+
+        BmpWriterRgb bmpWriterRgb = new BmpWriterRgb();
+        BufferedImage bufferedImage = new BufferedImage(2, 2, 5);
+        byte[] byteArray = bmpWriterRgb.getImageData(bufferedImage);
+
+        assertEquals( 24, bmpWriterRgb.getBitsPerPixel() );
+        assertEquals( 0, bmpWriterRgb.getPaletteSize() );
+        assertEquals( 16, byteArray.length );
+        assertEquals( "[0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]", Arrays.toString(byteArray) );
+
+    }
+
+
+}
