+++ src/test/java/org/apache/commons/imaging/formats/jpeg/iptc/IptcTypeLookupTest.java
@@ -14,15 +14,15 @@ import static org.junit.Assert.assertEquals;
 public class IptcTypeLookupTest{
 
 
-  @Test
-  public void testGetIptcTypeWithPositive() {
+    @Test
+    public void testGetIptcTypeWithPositive() {
 
-      IptcType iptcType = IptcTypeLookup.getIptcType(1117);
+        IptcType iptcType = IptcTypeLookup.getIptcType(1117);
 
-      assertEquals( 1117, iptcType.getType() );
-      assertEquals( "Unknown", iptcType.getName() );
+        assertEquals( 1117, iptcType.getType() );
+        assertEquals( "Unknown", iptcType.getName() );
 
-  }
+    }
 
 
 }
