+++ src/test/java/org/apache/commons/imaging/formats/jpeg/segments/JfifSegmentTest.java
