+++ src/test/java/org/apache/commons/imaging/formats/jpeg/segments/SosSegmentTest.java
