+++ src/test/java/org/apache/commons/imaging/formats/pnm/PbmFileInfoTest.java
