+++ src/test/java/org/apache/commons/imaging/formats/pnm/WhiteSpaceReaderTest.java
