+++ src/test/java/org/apache/commons/imaging/formats/psd/dataparsers/DataParserCmykTest.java
@@ -14,14 +14,14 @@ import static org.junit.Assert.assertEquals;
 public class DataParserCmykTest{
 
 
-  @Test
-  public void testGetBasicChannelsCount() {
+    @Test
+    public void testGetBasicChannelsCount() {
 
-      DataParserCmyk dataParserCmyk = new DataParserCmyk();
+        DataParserCmyk dataParserCmyk = new DataParserCmyk();
 
-      assertEquals(4, dataParserCmyk.getBasicChannelsCount());
+        assertEquals(4, dataParserCmyk.getBasicChannelsCount());
 
-  }
+    }
 
 
 }
