+++ src/test/java/org/apache/commons/imaging/formats/tiff/fieldtypes/FieldTypeLongTest.java
