+++ src/test/java/org/apache/commons/imaging/formats/tiff/taginfos/TagInfoByteTest.java
